+++ backend/canisters/user_index/impl/src/queries/http_request.rs
@@ -1,8 +1,9 @@
 use crate::{read_state, RuntimeState};
+use candid::Principal;
 use http_request::{build_json_response, encode_logs, extract_route, Route};
 use ic_cdk::query;
 use std::collections::BTreeMap;
-use types::{HttpRequest, HttpResponse, TimestampMillis};
+use types::{HttpRequest, HttpResponse, TimestampMillis, UserId};
 
 #[query]
 fn http_request(request: HttpRequest) -> HttpResponse {
@@ -16,6 +17,14 @@ fn http_request(request: HttpRequest) -> HttpResponse {
 
     fn get_metrics_impl(state: &RuntimeState) -> HttpResponse {
         build_json_response(&state.metrics())
+    }
+
+    fn get_user_metrics_impl(user_id: UserId, state: &RuntimeState) -> HttpResponse {
+        if let Some(metrics) = state.user_metrics(user_id) {
+            build_json_response(&metrics)
+        } else {
+            HttpResponse::not_found()
+        }
     }
 
     fn get_bot_users(state: &RuntimeState) -> HttpResponse {
@@ -46,6 +55,16 @@ fn http_request(request: HttpRequest) -> HttpResponse {
         Route::Metrics => read_state(get_metrics_impl),
         Route::Other(path, _) if path == "bots" => read_state(get_bot_users),
         Route::Other(path, _) if path == "new_users_per_day" => read_state(get_new_users_per_day),
+        Route::Other(path, _) => {
+            let parts: Vec<_> = path.split('/').collect();
+            if parts[0] == "usermetrics" {
+                let user_id: Option<UserId> = parts.get(1).and_then(|p| Principal::from_text(*p).ok()).map(|p| p.into());
+                if let Some(user_id) = user_id {
+                    return read_state(|state| get_user_metrics_impl(user_id, state));
+                }
+            }
+            HttpResponse::not_found()
+        }
         _ => HttpResponse::not_found(),
     }
 }